Combine intro into StartScreen
